src/models/image_recognition.py
"""
Image recognition module for identifying food items in fridge photos
"""
import cv2
import numpy as np
from PIL import Image

class FridgeImageAnalyzer:
    def __init__(self):
        # TODO: Load pre-trained food recognition model
        # For now, we'll use a placeholder
        self.model = None
        self.food_database = self._load_food_database()
    
    def _load_food_database(self):
        """Load database of known food items"""
        # This would typically load from a database or file
        food_items = {
            'tomato': {'name': 'Tomat', 'shelf_life': 7, 'carbon_footprint': 1.1},
            'banana': {'name': 'Pisang', 'shelf_life': 5, 'carbon_footprint': 0.9},
            'apple': {'name': 'Apel', 'shelf_life': 30, 'carbon_footprint': 0.5},
            'milk': {'name': 'Susu', 'shelf_life': 7, 'carbon_footprint': 1.5},
            'bread': {'name': 'Roti', 'shelf_life': 5, 'carbon_footprint': 1.0},
            'egg': {'name': 'Telur', 'shelf_life': 21, 'carbon_footprint': 0.8},
            'chicken': {'name': 'Ayam', 'shelf_life': 2, 'carbon_footprint': 3.2},
            'lettuce': {'name': 'Selada', 'shelf_life': 3, 'carbon_footprint': 0.3},
        }
        return food_items
    
    def analyze_image(self, image_path):
        """
        Analyze a fridge image and identify food items
        
        Args:
            image_path (str): Path to the fridge image
            
        Returns:
            list: List of identified food items with quantities
        """
        # Load image
        image = cv2.imread(image_path)
        if image is None:
            raise ValueError(f"Could not load image from {image_path}")
        
        # TODO: Implement actual food recognition using ML model
        # For now, we'll return sample data
        detected_items = self._mock_detection()
        
        return detected_items
    
    def _mock_detection(self):
        """Mock detection for demonstration purposes"""
        # In a real implementation, this would use a trained ML model
        # to detect food items in the image
        return [
            {'name': 'tomato', 'quantity': 3, 'days_until_expiry': 2},
            {'name': 'banana', 'quantity': 2, 'days_until_expiry': 1},
            {'name': 'bread', 'quantity': 1, 'days_until_expiry': 3},
            {'name': 'milk', 'quantity': 1, 'days_until_expiry': 5}
        ]
    
    def get_food_info(self, food_name):
        """Get information about a food item"""
        return self.food_database.get(food_name, None)

# Example usage
if __name__ == "__main__":
    analyzer = FridgeImageAnalyzer()
    # For testing without actual image
    items = analyzer._mock_detection()
    print("Detected items:")
    for item in items:
        info = analyzer.get_food_info(item['name'])
        print(f"- {info['name']}: {item['quantity']} items, {item['days_until_expiry']} days until expiry")