requirements.txt
pandas
numpy
prophet
opencv-python
tensorflow
Pillow
flask
scikit-learn
matplotlib
seaborn
requests
python-dotenv